+++ src/containers/TopicList.js
@@ -5,11 +5,12 @@ import TopicService from "../services/TopicService";
 
 export default class TopicList extends React.Component {
 
-    constructor(props){
+    constructor(props) {
         super(props);
-        this.state={
-            courseId:'', moduleId:'', lessonId:'', topics:[], topic:{title:''}
+        this.state = {
+            courseId: '', moduleId: '', lessonId: '', topics: [], topic: {title: ''}
         };
+
 
         this.setCourseId = this.setCourseId.bind(this);
         this.setModuleId = this.setModuleId.bind(this);
@@ -24,18 +25,17 @@ export default class TopicList extends React.Component {
         this.topicService = TopicService.instance;
     }
 
-    componentDidMount(){
-        this.setCourseId(this.props.match.params.courseId);
-        this.setModuleId(this.props.match.params.moduleId);
-        this.setLessonId(this.props.match.params.lessonId);
-        this.findAllTopicsForLesson(this.props.match.params.courseId, this.props.match.params.moduleId, this.props.match.params.lessonId);
-
+    componentDidMount() {
+        this.setCourseId(this.props.courseId);
+        this.setModuleId(this.props.moduleId);
+        this.setLessonId(this.props.lessonId);
     }
 
     componentWillReceiveProps(newProps) {
-        this.setCourseId(newProps.match.params.courseId);
-        this.setModuleId(newProps.match.params.moduleId);
-        this.setLessonId(newProps.match.params.lessonId);
+        this.setCourseId(newProps.courseId);
+        this.setModuleId(newProps.moduleId);
+        this.setLessonId(newProps.lessonId);
+        this.findAllTopicsForLesson(newProps.courseId, newProps.moduleId, newProps.lessonId);
     }
 
 
@@ -52,11 +52,14 @@ export default class TopicList extends React.Component {
     }
 
     setTopicTitle(event) {
-        this.setState({topic: {
+        this.setState({
+            topic: {
                 title: event.target.value
-            }})}
+            }
+        })
+    }
 
-    createTopic(){
+    createTopic() {
         this.topicService.createTopic
         (this.state.courseId, this.state.moduleId, this.state.lessonId, this.state.topic)
             .then(() => {
@@ -65,12 +68,11 @@ export default class TopicList extends React.Component {
             });
     }
 
-    renderListOfTopics(){
-        let topics = this.state.topics.map((topic) =>
-            {
-                return <TopicListItem  key={topic.id} topic={topic} delete={this.deleteTopic}
-                                        courseId={this.state.courseId} moduleId={this.state.moduleId}
-                                        lessonId={this.state.lessonId }/>
+    renderListOfTopics() {
+        let topics = this.state.topics.map((topic) => {
+                return <TopicListItem key={topic.id} topic={topic} delete={this.deleteTopic}
+                                      courseId={this.state.courseId} moduleId={this.state.moduleId}
+                                      lessonId={this.state.lessonId}/>
             }
         );
         return (
@@ -78,34 +80,48 @@ export default class TopicList extends React.Component {
         )
     }
 
-    deleteTopic(topicId){
-        this.topicService
-            .deleteTopic(topicId)
-            .then(() => {
-                this.findAllTopicsForLesson
-                (this.state.courseId, this.state.moduleId, this.state.lessonId)
-            });
+    deleteTopic(topicId) {
+        if (window.confirm('Are you sure you want to delete the selected topic ?')) {
+            this.topicService
+                .deleteTopic(topicId)
+                .then(() => {
+                    this.findAllTopicsForLesson
+                    (this.state.courseId, this.state.moduleId, this.state.lessonId)
+                });
+        }
     }
 
     findAllTopicsForLesson(courseId, moduleId, lessonId) {
         this.topicService
             .findAllTopicsForLesson(courseId, moduleId, lessonId)
-            .then((topics) => {this.setTopics(topics)});
+            .then((topics) => {
+                this.setTopics(topics)
+            });
     }
 
     setTopics(topics) {
         this.setState({topics: topics})
     }
 
-    render(){
-        return(
+    render() {
+        return (
             <div>
-            <input value={this.state.topic.title} onChange={this.setTopicTitle} placeholder="New Topic"/>
-            <button onClick= {this.createTopic} className="btn btn-primary btn-block">
-            <i className="fa fa-plus"></i></button>
-            <ul className="nav nav-pills">
-                {this.renderListOfTopics()}
-            </ul>
+                {/*<input value={this.state.topic.title} onChange={this.setTopicTitle} placeholder="New Topic"/>*/}
+                {/*<button onClick= {this.createTopic} className="btn btn-primary btn-block">*/}
+                {/*<i className="fa fa-plus"></i></button>*/}
+                <ul className="nav nav-pills">
+                    {this.renderListOfTopics()}
+                    <li className="nav-item cm-lessntab-li">
+                        <div className="input-group mb-3">
+                            <input type="text" value={this.state.topic.title} onChange={this.setTopicTitle}
+                                   placeholder="New Topic" className="form-control"/>
+                            <div className="input-group-append">
+                                <button onClick={this.createTopic} className="cm-add-button btn btn-primary float-right">
+                                    <i className="cm-add-icon fa fa-lg fa-plus"></i></button>
+                            </div>
+                        </div>
+                    </li>
+                </ul>
             </div>
         )
     }
